Track line numbers in Debian copyright files #13
This adds support for tracking line number when processing copyright
files. The approach is to only support this for copyright files and
keep a mapping of start/end line numbers by field name at the paragraph
level. This way existing fields do not need modifications and the core
code update is when paragraphs are created which is limited to a single
place.

There is also a new deb822 module replacing the email parser to parse
copyright files keep track of line numbers.

Signed-off-by: Philippe Ombredanne <pombredanne@nexb.com>